refactor(search): extract page-read + retrieval into shared modules
